feat: Init config for optional theming
- Cleanups
- Set init default APP
- Optional AppWithTheme WIP
