+++ README.md
@@ -1,9 +1,1 @@
-[![Node.js Package](https://github.com/NamicGreen/onboarding/actions/workflows/package_deployment.yml/badge.svg?branch=next&event=deployment_status)](https://github.com/NamicGreen/onboarding/actions/workflows/package_deployment.yml)
-
-[![Node.js Package](https://github.com/NamicGreen/onboarding/actions/workflows/package_deployment.yml/badge.svg?branch=next&event=registry_package)](https://github.com/NamicGreen/onboarding/actions/workflows/package_deployment.yml)
-
-[![Node.js Package](https://github.com/NamicGreen/onboarding/actions/workflows/package_deployment.yml/badge.svg?branch=next&event=release)](https://github.com/NamicGreen/onboarding/actions/workflows/package_deployment.yml)
-
-[![Node.js Package](https://github.com/NamicGreen/onboarding/actions/workflows/package_deployment.yml/badge.svg?branch=next&event=status)](https://github.com/NamicGreen/onboarding/actions/workflows/package_deployment.yml)
-
-[![Node.js Package](https://github.com/NamicGreen/onboarding/actions/workflows/package_deployment.yml/badge.svg?branch=next&event=workflow_run)](https://github.com/NamicGreen/onboarding/actions/workflows/package_deployment.yml)
+Comming soon
