+++ example/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@namicgreen/onboarding-example",
   "homepage": ".",
-  "version": "0.0.3",
+  "version": "0.0.4",
   "private": true,
   "scripts": {
     "start": "node ../node_modules/react-scripts/bin/react-scripts.js start",
