+++ example/src/App.test.tsx
@@ -1,9 +1,10 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import App from './App'
-
+/* import AppWithTheme from './AppWithTheme' */
 it('renders without crashing', () => {
   const div = document.createElement('div')
-  ReactDOM.render(<App />, div)
+	ReactDOM.render(<App />, div)
+/* 	ReactDOM.render(<AppWithTheme />, div) */
   ReactDOM.unmountComponentAtNode(div)
 })
