+++ example/src/index.tsx
@@ -3,7 +3,7 @@ import './index.css'
 import React from 'react'
 import ReactDOM from 'react-dom'
 import App from './App'
-
+/* import AppWithTheme from './AppWithTheme' */
 
 ReactDOM.render(<App />, document.getElementById('root'))
-
+/* ReactDOM.render(<AppWithTheme />, document.getElementById('root')) */
